+++ tests/integration/test_integration_orm.py
@@ -251,3 +251,35 @@ def test_every_field(Everything):
     assert record.link_count == 1
     assert record.lookup_error == [{"error": "#ERROR!"}]
     assert record.lookup_integer == [record.formula_integer]
+
+
+def test_attachments_upload(Everything, valid_img_url, tmp_path):
+    record: _Everything = Everything()
+    record.save()
+
+    # add an attachment via URL
+    record.attachments.append({"url": valid_img_url, "filename": "logo.png"})
+    record.save()
+    assert record.attachments[0]["url"] == valid_img_url
+
+    record.fetch()
+    assert record.attachments[0]["filename"] == "logo.png"
+    assert record.attachments[0]["type"] == "image/png"
+    assert record.attachments[0]["url"] != valid_img_url  # overwritten by Airtable
+
+    # add an attachment by uploading content
+    tmp_file = tmp_path / "sample.txt"
+    tmp_file.write_text("Hello, World!")
+    record.attachments.upload(tmp_file)
+    # ensure we got all attachments, not just the latest one
+    assert record.attachments[0]["filename"] == "logo.png"
+    assert record.attachments[0]["type"] == "image/png"
+    assert record.attachments[1]["filename"] == "sample.txt"
+    assert record.attachments[1]["type"] == "text/plain"
+
+    # ensure everything persists/loads correctly after fetch()
+    record.fetch()
+    assert record.attachments[0]["filename"] == "logo.png"
+    assert record.attachments[0]["type"] == "image/png"
+    assert record.attachments[1]["filename"] == "sample.txt"
+    assert record.attachments[1]["type"] == "text/plain"